Changed drawing flow. Added border around chart.

+++ app/src/main/java/com/dvinc/customchart/pointsgenerator/points/ChartPoint.kt
@@ -7,7 +7,7 @@ package com.dvinc.customchart.pointsgenerator.points
 
 interface ChartPoint {
 
-    fun getX(): Int
+    fun getX(): Float
 
-    fun getY(): Int
+    fun getY(): Float
 }
